Remove Mobility::Sequel::SQL namespace
